Fix bad type for txPower in PeriodicAdvertisingReport serialization
Bug: 69634768
Test: compilation
Merged-In: Icedfbaf1ba933637e935ada0fd98aea42c73f2b2

+++ core/java/android/bluetooth/le/PeriodicAdvertisingReport.java
@@ -71,7 +71,7 @@ public final class PeriodicAdvertisingReport implements Parcelable {
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeInt(syncHandle);
-        dest.writeLong(txPower);
+        dest.writeInt(txPower);
         dest.writeInt(rssi);
         dest.writeInt(dataStatus);
         if (data != null) {
